+++ src/Symfony/Bundle/FrameworkBundle/DependencyInjection/FrameworkExtension.php
@@ -2413,6 +2413,7 @@ class FrameworkExtension extends Extension
                 $container->getDefinition('notifier.channel.email')->setArgument(0, null);
             }
             $container->getDefinition('notifier.channel.sms')->setArgument(0, null);
+            $container->getDefinition('notifier.channel.push')->setArgument(0, null);
         }
 
         $container->getDefinition('notifier.channel_policy')->setArgument(0, $config['channel_policy']);